+++ resources/views/components/ui/context-menu-group.blade.php
@@ -1,3 +1,1 @@
-<div role="group" data-slot="context-menu-group" x-blat-labelledby="{ label: ':scope > [data-slot=context-menu-label]' }" {{ $attributes }}>
-    {{ $slot }}
-</div>
+<x-ui.menu-group :data-slot="'context-menu-group'" :label-slot="'context-menu-label'" {{ $attributes }}>{{ $slot }}</x-ui.menu-group>