+++ resources/views/components/ui/context-menu-separator.blade.php
@@ -1,1 +1,1 @@
-<div data-slot="context-menu-separator" role="separator" aria-orientation="horizontal" {{ $attributes->twMerge('bg-border -mx-1 my-1 h-px') }}></div>
+<x-ui.menu-separator :data-slot="'context-menu-separator'" {{ $attributes }} />