+++ resources/views/components/ui/dropdown-menu-label.blade.php
@@ -1,8 +1,7 @@
 @props(['inset' => false])
 
-<div
-    data-slot="dropdown-menu-label"
-    role="presentation"
-    @if ($inset) data-inset @endif
-    {{ $attributes->twMerge('px-2 py-1.5 text-sm font-medium data-[inset]:pl-8') }}
->{{ $slot }}</div>
+<x-ui.menu-label
+    :data-slot="'dropdown-menu-label'"
+    :inset="$inset"
+    {{ $attributes }}
+>{{ $slot }}</x-ui.menu-label>