+++ resources/views/components/ui/menubar-separator.blade.php
@@ -1,1 +1,1 @@
-<div data-slot="menubar-separator" role="separator" aria-orientation="horizontal" {{ $attributes->twMerge('bg-border -mx-1 my-1 h-px') }}></div>
+<x-ui.menu-separator :data-slot="'menubar-separator'" {{ $attributes }} />